+++ src/main/java/io/github/mastaan66/ledger/TransferDirection.java
@@ -1,0 +1,7 @@
+package io.github.mastaan66.ledger;
+
+public enum TransferDirection {
+    ALL,
+    INCOMING,
+    OUTGOING
+}